soc: fsl: rcpm: remove build warning
This removes build warning,
drivers/soc/fsl/rcpm.c: In function ‘rcpm_pm_prepare’:
drivers/soc/fsl/rcpm.c:126:37: warning: left shift count >= width of type [-Wshift-count-overflow]
          (u32)(((u64)(reg_offset[1] << (sizeof(u32) * 8) |
                                     ^~
drivers/soc/fsl/rcpm.c:131:38: warning: left shift count >= width of type [-Wshift-count-overflow]
           (u32)(((u64)(reg_offset[1] << (sizeof(u32) * 8) |

+++ drivers/soc/fsl/rcpm.c
@@ -123,12 +123,12 @@ static int rcpm_pm_prepare(struct device *dev)
 				}
 				/* Read value from register SCFG_SPARECR8 */
 				regmap_read(scfg_addr_regmap,
-					    (u32)(((u64)(reg_offset[1] << (sizeof(u32) * 8) |
+					    ((((u64)reg_offset[1] << (sizeof(u32) * 8) |
 					    reg_offset[2])) & 0xffffffff),
 					    &reg_value);
 				/* Write value to register SCFG_SPARECR8 */
 				regmap_write(scfg_addr_regmap,
-					     (u32)(((u64)(reg_offset[1] << (sizeof(u32) * 8) |
+					     ((((u64)reg_offset[1] << (sizeof(u32) * 8) |
 					     reg_offset[2])) & 0xffffffff),
 					     tmp | reg_value);
 			}
